+++ MarsExploration/Action.h
@@ -23,9 +23,9 @@ public:
 	void checkWaiting_P(Queue<Mission>& Polar, PriQ<Rover>* roverArray, int d);
 	void checkWaiting_M(HashTable<Mission>& Mountainous, Queue<int>& MountainousSort, PriQ<Rover>* roverArray, int d);
 
-	void MoveToExec_M(HashTable<Mission>& Mountainous, Queue<int>& MountainousSort, PriQ<Mission>& InExecution, int&Exec, int& Wait);
-	void MoveToExec_P(Queue<Mission>& Polar, PriQ<Mission>& InExecution, int& Exec, int& Wait);
-	void MoveToExec_E(PriQ<Mission>& Emergency, PriQ<Mission>& InExecution, int& Exec, int& Wait);
+	void MoveToExec_M(HashTable<Mission>& Mountainous, Queue<int>& MountainousSort, PriQ<Mission>& InExecution, float&Exec, float& Wait);
+	void MoveToExec_P(Queue<Mission>& Polar, PriQ<Mission>& InExecution, float& Exec, float& Wait);
+	void MoveToExec_E(PriQ<Mission>& Emergency, PriQ<Mission>& InExecution, float& Exec, float& Wait);
 	void AutoPromote(HashTable<Mission>& Mountainous, Queue<int>& MountainousSort, PriQ<Mission>& Emergency,int,int*);
 
 };
